feat(9pfs): Unify creation and deletion of new rootfs with mountRootfs
Unify the way we create/delete the monitor's rootfs when the mountRootfs
is set and we need to mount the container's rootfs as the guest's
rootfs. Hence, regardless of using block device or shared-fs, we will
separate the monitor's rootfs from the container's one and we can easily
then remove the new rootfs.

PR: https://github.com/urunc-dev/urunc/pull/194
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>