fix(unikraft): Update cli handling for unikraft
Change the way we handle cli arguments for Unikraft. The main reason for
this change is the way Unikraft handles arguments. In particular:

- In Qemu, argv[0] is always replaced byt the -kernel cli option of
Qemu. Even if it is empty
- In FC, argv[0] is set by the user, but in kraftkit the follow the same
convention as in Qemu.

However, here is the catch. Depending on the execution mode of Unikraft,
native or elfloader cli args are handled differently.
- In native execution, argv[0] does not really matter since it is
consumed by the application (which is part of the unikernel).
- In elfloader mode, argv[0] is consumed from the elfloader and
elfloader expects argv[1] to be the path to the application binary.

To make it more clear, let's take as an exmaple redis. If we want to
specify the path to the redis config we can pass it as argument in the
redis binary (e.g. redis-server /conf/redis.conf).
- In native execution, we would set argv[0] as "redis-server" and argv[1]
- In elfloader execution, we need set argv[0] as anything dummy (e.g. "app")
and argv[1] as "/pat/to/redis-server" while argv[2] as "/conf/redis.conf"

Taking into consideration all the above, this commit always sets
argv[0] as "Unikraft". This is similar to Kraftkit, which always sets
arv[0] as the kernel to boot.

Also remove environment variable setting, if we have an old versionn of
Unikraft, since older versions do not support environment variables.

PR: https://github.com/urunc-dev/urunc/pull/195
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>